Create Data Transfer Objects
Creating DTOs to decouple service layer and database layer. Allows us to guarantee what we are getting in the request and responses to the db layer and client side. Also allows us to manipulate the payload as needed (can omit properties, make it smaller, make it more legible for client etc.)